+++ src/main/java/com/jpa/commerce/entity/order/OrderEntity.java
@@ -1,4 +1,6 @@
-package com.jpa.commerce.entity.orders;
+package com.jpa.commerce.entity.order;
+
+import java.sql.Date;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -9,21 +11,28 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
+import org.hibernate.annotations.ColumnDefault;
+
 import com.jpa.commerce.entity.common.CommonEntity;
+import com.jpa.commerce.entity.product.ProductsEntity;
 
 @Entity
-@Table(name = "OrderStatuses")
-public class OrderStatusesEntity extends CommonEntity {
+@Table(name = "orders")
+public class OrderEntity extends CommonEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne
-    @JoinColumn(name = "order_id")
-    private OrderEntity orders;
+    @JoinColumn(name = "product_id")
+    private ProductsEntity product;
 
     @Column
-    private String status;
+    @ColumnDefault("0")
+    private int quantity;
+
+    @Column
+    private Date orderDate;
 
 }
